pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vdbbench"
version = "0.1.0"
description = "A benchmarking tool for vector databases"
readme = "README.md"
license = "Apache-2.0"

authors = [
    {name = "Your Name", email = "your.email@example.com"}
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Topic :: Database :: Database Engines/Servers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "pandas",
    "matplotlib",
    "pymilvus",
    "qdrant-client",
    "pinecone-client",
    "tqdm",
    "seaborn",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
    "mypy",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/vdbbench"
"Bug Tracker" = "https://github.com/yourusername/vdbbench/issues"

[project.scripts]
vdbbench = "vector_benchmark.cli:main"

[tool.setuptools]
package-dir = {"" = "."}
packages = ["vector_benchmark"]

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"